// webview/src/shared/types.ts
export enum ProjectManagementTool {
  Linear = 'Linear',
  Asana = 'Asana',
  Shortcut = 'Shortcut',
  AzureDevOps = 'AzureDevOps',
  GitHub = 'GitHub',
  Jira = 'Jira',
}